Updated MainPage.tsx

+++ src/pages/MainPage.tsx
@@ -50,7 +50,7 @@ export default function MainPage() {
 
   const formatDate = (dateString: string) => {
     const date = new Date(dateString);
-    return date.toLocaleDateString('en-US', { year: 'numeric', month: 'long', day: 'numeric' });
+    return date.toLocaleDateString('en-US', { year: 'numeric', month: 'short', day: 'numeric' });
   };
 
   const handlePhoneChange = async (phone: string) => {
